# swagger/api-docs.yaml
openapi: 3.0.0
info:
  title: User Authentication API
  version: 1.0.0
  description: API for user login and logout functionality
servers:
  - url: http://localhost:3000/api # Replace with your actual base API URL
...
paths:
...
  /messages:
    post:
      tags:
        - Messages
      summary: Create a new message
      description: Creates a new message in a conversation.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                conversationId:
                  type: integer
                  description: Conversation's ID
                senderId:
                  type: integer
                  description: Sender's ID
                message:
                  type: string
                  description: Text message
                messageImage:
                  type: string
                  format: binary
                  description: Message's Image
      responses:
        201:
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
  /messages/{id}:
    get:
      tags:
        - Messages
      summary: Get messages of a specific conversation
      description: Retrieves an array of messages of an conversation with conversationId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Conversation ID
      responses:
        200:
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/Message"
        404:
          description: Conversation not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Conversation not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

    put:
      tags:
        - Messages
      summary: Create a new message
      description: Creates a new message in a conversation.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Message ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                conversationId:
                  type: integer
                  description: Conversation's ID
                senderId:
                  type: integer
                  description: Sender's ID
                message:
                  type: string
                  description: Text message
                messageImage:
                  type: string
                  format: binary
                  description: Message's Image
      responses:
        201:
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: Message not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Message not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

    delete:
      tags:
        - Messages
      summary: Delete a message
      description: Delete a message.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Message ID
      responses:
        200:
          description: Friend deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friend"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: Message not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Friend:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        friendId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        message:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        imageUrl:
          type: string
        createdAt:
          format: date-time
        updatedAt:
          format: date-time